一些尝试

+++ src/main/java/com/memory/interfaces/Kernel32_DLL.java
@@ -1,11 +1,9 @@
 package com.memory.interfaces;
 
-import com.memory.structure.MEMORY_BASIC_INFORMATION;
-import com.memory.structure.MODULEENTRY32;
-import com.memory.structure.PROCESSENTRY32;
-import com.memory.structure.SYSTEM_INFO;
+import com.memory.structure.*;
 import com.sun.jna.Native;
 import com.sun.jna.Pointer;
+import com.sun.jna.platform.win32.Tlhelp32;
 import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.win32.StdCallLibrary;
 import com.sun.jna.win32.W32APIOptions;
@@ -210,7 +208,7 @@ public interface Kernel32_DLL extends StdCallLibrary {
      * 函数具体详情,请参阅MSDN文档
      * http://msdn.microsoft.com/zh-cn/library/aa366899%28v=vs.85%29.aspx
      **/
-    boolean VirtualProtectEx(int hProcess, int lpAddress, int dwSize, int flNewProtect, int lpflOldProtect);
+    boolean VirtualProtectEx(long hProcess, long lpAddress, int dwSize, int flNewProtect, int lpflOldProtect);
 
     /**
      * 获取当前系统的信息
